OCPBUGS-48642: Remove `network-attachment-definition-plugin`
After some further investigation it seems like none of the code is used in this package, aside from a k8s model and kind, which have been rehomed to `@console/app`.
