+++ frontend/packages/console-app/package.json
@@ -17,7 +17,6 @@
     "@console/knative-plugin": "0.0.0-fixed",
     "@console/local-storage-operator-plugin": "0.0.0-fixed",
     "@console/metal3-plugin": "0.0.0-fixed",
-    "@console/network-attachment-definition-plugin": "0.0.0-fixed",
     "@console/operator-lifecycle-manager": "0.0.0-fixed",
     "@console/operator-lifecycle-manager-v1": "0.0.0-fixed",
     "@console/patternfly": "0.0.0-fixed",
