+++ frontend/packages/console-app/src/models/index.ts
@@ -39,3 +39,17 @@ export const EndPointSliceModel: K8sKind = {
   namespaced: true,
   plural: 'endpointslices',
 };
+
+export const NetworkAttachmentDefinitionModel: K8sKind = {
+  label: 'Network Attachment Definition',
+  labelPlural: 'Network Attachment Definitions',
+  apiVersion: 'v1',
+  apiGroup: 'k8s.cni.cncf.io',
+  plural: 'network-attachment-definitions',
+  namespaced: true,
+  abbr: 'NAD',
+  kind: 'NetworkAttachmentDefinition',
+  id: 'network-attachment-definition',
+  crd: true,
+  legacyPluralURL: true,
+};
