--- frontend/packages/network-attachment-definition-plugin/OWNERS
@@ -1,10 +1,0 @@
-reviewers:
-  - pcbailey
-  - mareklibra
-  - rawagner
-approvers:
-  - pcbailey
-  - mareklibra
-  - rawagner
-labels:
-  - component/network-attachment-definition
