--- frontend/packages/network-attachment-definition-plugin/console-extensions.json
@@ -1,1 +1,0 @@
-[]
