--- frontend/packages/network-attachment-definition-plugin/locales/OWNERS
@@ -1,2 +1,0 @@
-labels:
-  - kind/i18n
