--- frontend/packages/network-attachment-definition-plugin/locales/en/kubevirt-plugin.json
@@ -1,32 +1,0 @@
-{
-  "No {{label}} found": "No {{label}} found",
-  "Create {{label}}": "Create {{label}}",
-  "Learn how to use {{label}}": "Learn how to use {{label}}",
-  "NetworkAttachmentDefinition details": "NetworkAttachmentDefinition details",
-  "Type": "Type",
-  "Details": "Details",
-  "Network attachment definition name cannot be empty": "Network attachment definition name cannot be empty",
-  "Network attachment definition name can contain only alphanumeric characters": "Network attachment definition name can contain only alphanumeric characters",
-  "Network attachment definition name must start/end with alphanumeric character": "Network attachment definition name must start/end with alphanumeric character",
-  "Network attachment definition name cannot contain uppercase characters": "Network attachment definition name cannot contain uppercase characters",
-  "Network attachment definition name is too long": "Network attachment definition name is too long",
-  "Network attachment definition name is too short": "Network attachment definition name is too short",
-  "Network Type": "Network Type",
-  "Edit YAML": "Edit YAML",
-  "Networks are not project-bound. Using the same name creates a shared NAD.": "Networks are not project-bound. Using the same name creates a shared NAD.",
-  "Create network attachment definition": "Create network attachment definition",
-  "Name": "Name",
-  "Description": "Description",
-  "<0>OpenShift Virtualization Operator</0> or <3>SR-IOV Network Operator </3>needs to be installed on the cluster, in order to pick the Network Type.": "<0>OpenShift Virtualization Operator</0> or <3>SR-IOV Network Operator </3>needs to be installed on the cluster, in order to pick the Network Type.",
-  "Create": "Create",
-  "Cancel": "Cancel",
-  "Resource name": "Resource name",
-  "VLAN tag number": "VLAN tag number",
-  "IP address management": "IP address management",
-  "Bridge name": "Bridge name",
-  "MAC spoof check": "MAC spoof check",
-  "Bridge mapping": "Bridge mapping",
-  "Physical network name. A bridge mapping must be configured on cluster nodes to map between physical network names and Open vSwitch bridges.": "Physical network name. A bridge mapping must be configured on cluster nodes to map between physical network names and Open vSwitch bridges.",
-  "MTU": "MTU",
-  "VLAN": "VLAN"
-}
