--- frontend/packages/network-attachment-definition-plugin/locales/es/network-attachment-definition-plugin.json
@@ -1,36 +1,0 @@
-{
-  "No {{label}} found": "No se encontró {{label}}",
-  "Create {{label}}": "Crear {{label}}",
-  "Learn how to use {{label}}": "Aprenda a utilizar {{label}}",
-  "NetworkAttachmentDefinition details": "Detalles de NetworkAttachmentDefinition",
-  "Type": "Tipo",
-  "Details": "Detalles",
-  "Network attachment definition name cannot be empty": "El nombre de la definición del archivo adjunto de red no puede estar vacío",
-  "Network attachment definition name can contain only alphanumeric characters": "El nombre de la definición del archivo adjunto de red puede contener solo caracteres alfanuméricos",
-  "Network attachment definition name must start/end with alphanumeric character": "El nombre de la definición del archivo adjunto de red debe comenzar o terminar con un carácter alfanumérico",
-  "Network attachment definition name cannot contain uppercase characters": "El nombre de la definición del archivo adjunto de red no puede contener caracteres en mayúsculas",
-  "Network attachment definition name is too long": "El nombre de la definición del archivo adjunto de red es demasiado largo",
-  "Network attachment definition name is too short": "El nombre de la definición del archivo adjunto de red es demasiado corto",
-  "Network Type": "Tipo de red",
-  "Edit YAML": "Editar YAML",
-  "Networks are not project-bound. Using the same name creates a shared NAD.": "Las redes no están vinculadas a proyectos. Usar el mismo nombre crea un NAD compartido.",
-  "Create network attachment definition": "Crear definición de archivo adjunto de red",
-  "Name": "Nombre",
-  "Description": "Descripción",
-  "<0>OpenShift Virtualization Operator</0> or <3>SR-IOV Network Operator </3>needs to be installed on the cluster, in order to pick the Network Type.": "El <0>Operador de virtualización OpenShift</0> u <3>Operador de red SR-IOV</3> deben instalarse en el clúster para poder elegir el tipo de red.",
-  "Create": "Crear",
-  "Cancel": "Cancelar",
-  "Assigns IP addresses from this subnet to Pods and VirtualMachines.": "Asigna direcciones IP de esta subred a pods y máquinas virtuales.",
-  "Tech preview": "Vista previa de tecnología",
-  "Resource name": "Nombre del recurso",
-  "VLAN tag number": "Número de etiqueta VLAN",
-  "IP address management": "Gestión de direcciones IP",
-  "Bridge name": "Nombre del puente",
-  "MAC spoof check": "Comprobación de falsificación de MAC",
-  "Bridge mapping": "Mapeo de puentes",
-  "Physical network name. Bridge mapping must be configured on cluster nodes to map between physical network names and Open vSwitch bridges.": "Nombre de la red física. Se debe configurar una asignación de puentes en los nodos del clúster para asignar entre los nombres de redes físicas y los puentes Open vSwitch.",
-  "MTU": "MTU",
-  "VLAN": "VLAN",
-  "Invalid subnet format": "Formato de subred no válido",
-  "Invalid IP address or subnet format": "Dirección IP o formato de subred no válidos"
-}
