--- frontend/packages/network-attachment-definition-plugin/locales/fr/kubevirt-plugin.json
@@ -1,32 +1,0 @@
-{
-  "No {{label}} found": "{{label}} introuvable",
-  "Create {{label}}": "Créer {{label}}",
-  "Learn how to use {{label}}": "Apprenez comment utiliser {{label}}",
-  "NetworkAttachmentDefinition details": "Détails de la définition de rattachement de réseau",
-  "Type": "Type",
-  "Details": "Détails",
-  "Network attachment definition name cannot be empty": "Le nom de la définition de rattachement de réseau ne peut pas être vide.",
-  "Network attachment definition name can contain only alphanumeric characters": "Le nom de la définition de rattachement de réseau ne peut contenir que des caractères alphanumériques.",
-  "Network attachment definition name must start/end with alphanumeric character": "Le nom de la définition de rattachement de réseau doit commencer et se terminer par un caractère alphanumérique.",
-  "Network attachment definition name cannot contain uppercase characters": "Le nom de la définition de rattachement de réseau ne peut pas contenir de caractères majuscules.",
-  "Network attachment definition name is too long": "Le nom de la définition de rattachement de réseau est trop long.",
-  "Network attachment definition name is too short": "Le nom de la définition de rattachement de réseau est trop court.",
-  "Network Type": "Type de réseau",
-  "Edit YAML": "Modifier YAML",
-  "Networks are not project-bound. Using the same name creates a shared NAD.": "Les réseaux ne sont pas liés à un projet. L’utilisation du même nom crée un NAD partagé.",
-  "Create network attachment definition": "Créer une définition de rattachement de réseau",
-  "Name": "Nom",
-  "Description": "Description",
-  "<0>OpenShift Virtualization Operator</0> or <3>SR-IOV Network Operator </3>needs to be installed on the cluster, in order to pick the Network Type.": "L’<0>opérateur de virtualisation OpenShift</0> ou l’<3>opérateur réseau SR-IOV</3> doit être installé sur le cluster, afin de choisir le type de réseau.",
-  "Create": "Créer",
-  "Cancel": "Annuler",
-  "Resource name": "Nom de la ressource",
-  "VLAN tag number": "Numéro de balise VLAN",
-  "IP address management": "Gestion des adresses IP",
-  "Bridge name": "Nom du pont",
-  "MAC spoof check": "Vérification d’usurpation d’identité MAC",
-  "Bridge mapping": "Mappage de ponts",
-  "Physical network name. A bridge mapping must be configured on cluster nodes to map between physical network names and Open vSwitch bridges.": "Nom du réseau physique. Un mappage de ponts doit être configuré sur les nœuds du cluster pour mapper les noms de réseaux physiques et les ponts Open vSwitch.",
-  "MTU": "MTU",
-  "VLAN": "VLAN"
-}
