--- frontend/packages/network-attachment-definition-plugin/locales/ja/kubevirt-plugin.json
@@ -1,32 +1,0 @@
-{
-  "No {{label}} found": "{{label}} が見つかりません",
-  "Create {{label}}": "{{label}} の作成",
-  "Learn how to use {{label}}": "{{label}} の使用方法の詳細",
-  "NetworkAttachmentDefinition details": "NetworkAttachmentDefinition の詳細",
-  "Type": "タイプ",
-  "Details": "詳細",
-  "Network attachment definition name cannot be empty": "ネットワーク接続定義名は空にできません",
-  "Network attachment definition name can contain only alphanumeric characters": "ネットワーク接続定義名には、英数字のみを含めることができます",
-  "Network attachment definition name must start/end with alphanumeric character": "ネットワーク接続定義名は英数字で開始/終了する必要があります",
-  "Network attachment definition name cannot contain uppercase characters": "ネットワーク接続定義名に大文字は使用できません",
-  "Network attachment definition name is too long": "ネットワーク接続定義名が長すぎます",
-  "Network attachment definition name is too short": "ネットワーク接続定義名が短すぎます",
-  "Network Type": "ネットワークタイプ",
-  "Edit YAML": "YAML の編集",
-  "Networks are not project-bound. Using the same name creates a shared NAD.": "ネットワークはプロジェクトにバインドされていません。同じ名前を使用すると共有 NAD が作成されます。",
-  "Create network attachment definition": "ネットワーク接続定義の作成",
-  "Name": "名前",
-  "Description": "説明",
-  "<0>OpenShift Virtualization Operator</0> or <3>SR-IOV Network Operator </3>needs to be installed on the cluster, in order to pick the Network Type.": "ネットワークタイプを選択するには、<0>OpenShift Virtualization Operator</0> または <3>SR-IOV Network Operator</3> がクラスターにインストールされている必要があります。",
-  "Create": "作成",
-  "Cancel": "キャンセル",
-  "Resource name": "リソース名",
-  "VLAN tag number": "VLAN タグ番号",
-  "IP address management": "IP アドレス管理",
-  "Bridge name": "ブリッジ名",
-  "MAC spoof check": "MAC spoof チェック",
-  "Bridge mapping": "ブリッジマッピング",
-  "Physical network name. A bridge mapping must be configured on cluster nodes to map between physical network names and Open vSwitch bridges.": "物理ネットワーク名。物理ネットワーク名と Open vSwitch ブリッジの間をマッピングするには、クラスターノード上にブリッジマッピングを設定する必要があります。",
-  "MTU": "MTU",
-  "VLAN": "VLAN"
-}
