--- frontend/packages/network-attachment-definition-plugin/locales/zh/network-attachment-definition-plugin.json
@@ -1,36 +1,0 @@
-{
-  "No {{label}} found": "没有找到 {{label}}",
-  "Create {{label}}": "创建{{label}}",
-  "Learn how to use {{label}}": "了解如何使用 {{label}}",
-  "NetworkAttachmentDefinition details": "网络附加定义详情",
-  "Type": "类型",
-  "Details": "详情",
-  "Network attachment definition name cannot be empty": "网络附加定义名不能为空",
-  "Network attachment definition name can contain only alphanumeric characters": "网络附加定义名只能包含字母数字字符",
-  "Network attachment definition name must start/end with alphanumeric character": "网络附加定义名必须使用字母数字字符开始和结束",
-  "Network attachment definition name cannot contain uppercase characters": "网络附加定义名不能包含大写字符",
-  "Network attachment definition name is too long": "网络附加定义名太长",
-  "Network attachment definition name is too short": "网络附加定义名太短",
-  "Network Type": "网络类型",
-  "Edit YAML": "编辑 YAML",
-  "Networks are not project-bound. Using the same name creates a shared NAD.": "网络不是与项目绑定的。使用相同的名称会创建一个共享的 NAD。",
-  "Create network attachment definition": "创建网络附加定义",
-  "Name": "名称",
-  "Description": "描述",
-  "<0>OpenShift Virtualization Operator</0> or <3>SR-IOV Network Operator </3>needs to be installed on the cluster, in order to pick the Network Type.": "<0>OpenShift Virtualization Operator</0> 或 <3>SR-IOV Network Operator</3> 需要在集群中安装，以便选择网络类型。",
-  "Create": "创建",
-  "Cancel": "取消",
-  "Assigns IP addresses from this subnet to Pods and VirtualMachines.": "将这个子网的 IP 地址分配给 Pod 和 VirtualMachines。",
-  "Tech preview": "技术预览",
-  "Resource name": "资源名称",
-  "VLAN tag number": "VLAN 标签号",
-  "IP address management": "IP 地址管理",
-  "Bridge name": "网桥名称",
-  "MAC spoof check": "MAC spoof 检查",
-  "Bridge mapping": "网桥映射",
-  "Physical network name. Bridge mapping must be configured on cluster nodes to map between physical network names and Open vSwitch bridges.": "物理网络名称。必须在集群节点上配置网桥映射，以便在物理网络名称和 Open vSwitch 网桥之间进行映射。",
-  "MTU": "MTU",
-  "VLAN": "VLAN",
-  "Invalid subnet format": "无效的子网格式",
-  "Invalid IP address or subnet format": "无效的 IP 地址或子网格式"
-}
