--- frontend/packages/network-attachment-definition-plugin/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@console/network-attachment-definition-plugin",
-  "version": "0.0.0-fixed",
-  "description": "Network attachment definition add-on",
-  "private": true,
-  "main": "src/index.ts",
-  "scripts": {
-    "lint": "yarn --cwd ../.. eslint packages/network-attachment-definition-plugin"
-  },
-  "dependencies": {
-    "@console/plugin-sdk": "0.0.0-fixed",
-    "@console/shared": "0.0.0-fixed",
-    "@console/internal": "0.0.0-fixed",
-    "unique-names-generator": "4.3.1"
-  },
-  "consolePlugin": {
-    "entry": "src/plugin.ts"
-  }
-}
