--- frontend/packages/network-attachment-definition-plugin/src/components/SubnetHelperText/SubnetHelperText.scss
@@ -1,10 +1,0 @@
-.kv-subnets-container {
-  &__info-icon {
-    color: var(--pf-t--color--blue--40);
-    margin-right: var(--pf-t--global--spacer--sm);
-  }
-  &__helper-text {
-    font-weight: bolder;
-    color: var(--pf-t--global--text--color--link--default);
-  }
-}
