--- frontend/packages/network-attachment-definition-plugin/src/components/TechPreview/TechPreview.scss
@@ -1,6 +1,0 @@
-.kv-tech-preview-label {
-  color: var(--pf-v6-c-button--m-danger--Color);
-  background-color: var(--pf-v6-c-button--m-danger--BackgroundColor);
-  margin-left: var(--pf-t--global--spacer--md);
-  padding: 0 var(--pf-t--global--spacer--sm);
-}
