--- frontend/packages/network-attachment-definition-plugin/src/components/network-attachment-definitions/NetworkAttachmentDefinition.scss
@@ -1,15 +1,0 @@
-.nad-quickstart-icon {
-  margin-right: var(--pf-t--global--spacer--sm);
-}
-
-.kv-nad-form-field--spacer {
-  margin-top: var(--pf-t--global--spacer--sm);
-}
-
-.kv-nad-form-checkbox--alignment {
-  height: 100%;
-}
-
-.network-type-options--help-icon {
-  cursor: pointer;
-}
