--- frontend/packages/network-attachment-definition-plugin/src/components/network-attachment-definitions/index.ts
@@ -1,3 +1,0 @@
-export * from './NetworkAttachmentDefinition';
-export * from './NetworkAttachmentDefinitionDetails';
-export * from './NetworkAttachmentDefinitionDetailsPage';
