--- frontend/packages/network-attachment-definition-plugin/src/components/network-attachment-definitions/types.ts
@@ -1,15 +1,0 @@
-import { NetworkAttachmentDefinitionConfig } from '../../types';
-
-export type NetAttachDefBundle = {
-  name?: string;
-  namespace?: string;
-  type?: string;
-  metadata?: any;
-  configJSON?: NetworkAttachmentDefinitionConfig;
-  netAttachDef: any;
-};
-
-export type NetworkAttachmentDefinitionsPageProps = {
-  filterLabel: string;
-  namespace?: string;
-};
