--- frontend/packages/network-attachment-definition-plugin/src/index.ts
@@ -1,1 +1,0 @@
-export * from './models';
