--- frontend/packages/network-attachment-definition-plugin/src/plugin.ts
@@ -1,3 +1,0 @@
-const plugin = [];
-
-export default plugin;
