--- frontend/packages/network-attachment-definition-plugin/src/types/index.ts
@@ -1,1 +1,0 @@
-export * from './types';
