--- frontend/packages/network-attachment-definition-plugin/src/types/types.ts
@@ -1,53 +1,0 @@
-import { K8sResourceKind } from '@console/internal/module/k8s';
-
-export type NetworkAttachmentDefinitionAnnotations = {
-  description?: string;
-  'k8s.v1.cni.cncf.io/resourceName'?: string;
-};
-
-export type IPAMConfig = {
-  type?: string;
-  subnet?: string;
-  dataDir?: string;
-};
-
-export type NetworkAttachmentDefinitionPlugin = {
-  [key: string]: any;
-};
-
-export type NetworkAttachmentDefinitionConfig = {
-  cniVersion: string;
-  name: string;
-  type?: string;
-  bridge?: string;
-  isGateway?: true;
-  macspoofchk?: boolean;
-  vlan?: number;
-  ipam?: IPAMConfig;
-  plugins?: NetworkAttachmentDefinitionPlugin[];
-  topology?: string;
-  netAttachDefName?: string;
-  preserveDefaultVlan?: boolean;
-  vlanID?: number;
-  mtu?: number;
-  subnets?: string;
-  excludeSubnets?: string;
-};
-
-// The config is a JSON object with the NetworkAttachmentDefinitionConfig type stored as a string
-export type NetworkAttachmentDefinitionSpec = {
-  config: string;
-};
-
-export type NetworkAttachmentDefinitionKind = {
-  spec?: NetworkAttachmentDefinitionSpec;
-} & K8sResourceKind;
-
-export type TypeParamsDataItem = {
-  value?: any;
-  validationMsg?: string;
-};
-
-export type TypeParamsData = {
-  [key: string]: TypeParamsDataItem;
-};
