+++ frontend/webpack.config.ts
@@ -298,10 +298,6 @@ const config: Configuration = {
         { from: path.resolve(__dirname, './packages/gitops-plugin/locales'), to: 'locales' },
         { from: path.resolve(__dirname, './packages/metal3-plugin/locales'), to: 'locales' },
         { from: path.resolve(__dirname, './packages/vsphere-plugin/locales'), to: 'locales' },
-        {
-          from: path.resolve(__dirname, './packages/network-attachment-definition-plugin/locales'),
-          to: 'locales',
-        },
         { from: path.resolve(__dirname, './packages/patternfly/locales'), to: 'locales' },
         { from: path.resolve(__dirname, './packages/insights-plugin/locales'), to: 'locales' },
         {
